CHANGE:Started using the Functional API
The subclasing models do not allow multiple inputs. Therefore started
the use of the Functional API which uses the DAG architecture for
building models that can take in multiple inputs. This provides one with
the necessary versatility to use both image data and metadata at the
same time. Still requires changes to the other testing models.