NEW:Started the second iteration of the GUI
The data could not be transferred from class to class using the more
commonly structured GUI app methods. Therefore the entire GUI was
restructured to be easier to transfer data to different frames. Decided
to use functions for the frames and one class for the container and
initialized all frames with the ability to switch to the next page based
on other functions using the tkraise function at the end of the
function.

The first iteration will be reviewed and the Tk window may be used to
store data instead. This may be more difficult, but will allow better
modularity.