CHANGE: Several Functions Were Added.
Added a function to convert the dicom pixel arrays into images. This was
done due to the difficulty of loading the raw pixel arrays. There is not
enough memory to load the data on to a tensorflow dataset.

Made corrections to the function that extracts key images. This
function has now been generalized to extract the paths of any image
dataset in any order.

Modified the function for extracting data from the DICOM file. This
function now simply extracts the data and organizes it into a
dictionary. This function can be used in tune with a for loop or the map
function to develop a list of dictionaries (that can be passed through
pandas DataFrame and create a df variable).

Created a function to transform the data. This function works to change
any non-categorical variable into its categorical version. This allows
the data to be transformed while in its dictionary form. Alternatively,
pandas DataFrame functionality can be utilized instead to transform
data.