CHANGE: Started Working on training the data
Finally achieved the training phase of the model. Currently encountering
an issue where the zeroth dimension of the machine learning model does
not match at the concatenation layer. This may have to be fixed at the
dense layer or the end of the image processing layer. May be resolved by
adding a new layer that mirrors the layer in relation to the categorical
input.