CHANGE: Started Full Training of models
Specified the batch size of the input data to fix the batch size itself
while removing dense layers after the Flatten layer. Although the model
is now fully capable of training itself, it is still not able to save
itself. Further research is required for saving the machine learning
model.