CHANGE:Developed new model
Made corrections to the data which allowed me to observe the increase in
training. Currently encountering an issue where the data does not allow
for better accuracy between models. For this reason the RMSprop
optimizer is being used at the moment. The next possible changes in
design may be attributed to the model architecture now that I have
discovered how to prevent the padding from occuring between
convolutions.