CHANGE:Created the base models for comparison
Created a set of base models meant to be used for comparing performance
between the base models and the finalized models. The base models so far
are two, one model that does not include the factor of age into the
calculation and another that does include age into the factors, but does
not use CNN architecture.