CHANGE: Addition of functions
Added the predict function from the capapp library to the pipeline. This
change  was made while keeping in mind that the predict function feeds
data into a model and loads the results on to a DataFrame.

Added the load_testing_data function and changed the load_data2 function
to the load_training_data function as both have different ways to load
the data for the model. The training data requires a dictionary with
key:value pairs where the values are lists while the test data requires
a list of dictionaries to create a DataFrame and feed that into the
model for predictions.

Also added the python syntax for knowing the result of the function
while minimized for most functions.