CHANGE:Parameters for the final model
Changed the first Convolutional Layer filters to 1200 and added a higher
dropout rate for the layer. The purpose of this is to avoid the incident
of overfitting.

While testing the finalized model, I noticed that the majority of the
predictions had probabilities close to 50% (meaning that the model may
be flipping a coin to predict whether the tumor is benign or malignant).
Changed the parameters to better fit this, only requirement left is to
train the model and retest.