CHANGE:Developed Functions further
Added conditional statements to the load_training_data function for the
sake of data validation. Also added a separate variable that will allow
for faster loading of the training data by removing the data balancing
step. This is much faster as the first training session creates and
downloads the balanced data while the rest of the other sessions load
the saved trained data.

Created a function that works to rescale the images. This function is
meant to make abstract the part of the algorithm that is already in use
from a long time ago in multiple places.