NEW:Function for data balancing
Developed a new function for balancing the data set. Added a new factor
for balancing the data (abnormality) which can be a large factor in
influencing the accuracy of the model as calcification can be very
similar to tumors.

Started the removal of the try except statements for the extract data
function.