NEW:Started using the pytest test framework.
Changed from the unittest to pytest as pytest was the most used python
testing framework. This framework also interfaces with other testing
frameworks, such as python's unittest and or the nose framework. The
main purpose of using pytest is to get familiar with a popular framework
that may be used throughout a job and develop test-driven development
skills.

The next step in the framework is to develop tests for the rest of the
functions used throughout the pipeline (mainly involving the loading
features of the pipeline such as rescaling images, data balancing, and
calculating the confusion matrix).