CHANGE: Replace hardcoding to for loop in function
Redeveloped the balance_data function to use a for loop to balance the
data set. Added parameter that will allow the user to determine the
sample size and placed a conditional statement in the case that the
sample size per group is larger than the size of the group itself. In
the case that the sample size per group is larger, the algorithm will
use the max size of the group.