CHANGE: Minor alterations to the module
Changed the batch size variable to allow for changes in batch size based
on the sample size chosen. Selected a sample size of 1,500 after
developing the changes to the function for loading the training data.

Added history functionality to observe the training variable together
with the AUC score, accuracy, and loss. decreased the number of epochs
to the point where the AUC score reaches approximately 1.00.