CHANGE: Developed function to test model accuracy.
Developed a function for testing the accuracy of the models. Changed the
model output fixture function to sample 1,000 subjects instead of using
only 1 sample. Completely changed how the model output was tested after
experimenting, this new algorithm allows one to make sure that the
outputs sum up to 1.0 (this is because they are meant to be
probabilities).