CHANGE: Fixed testing.
Made major corrections regarding the use of the load_testing_data
function. Allowed the function to output a singular DataFrame as opposed
to two different DataFrames by pre merging the DataFrame to be used for
model testing and the DataFrame containing the actual classifications.
Added further documentation to the function.

Corrected conditional statement for the predict function. Added further
documentation to the function.

Changed the calculate_confusion_matrix function to handle only one
input.