CHANGE: Fixed duplicates appearing when testing
Fixed bug where the pandas merging feature used within the
load_testing_data function caused duplicates to be created after
extracting the data required to test the model. This was done by
creating a dictionary out of the iterable row and using the dictionary's
built in update function to update the data into a new DataFrame.